+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/ColorimetryLiquidActivity.java
@@ -443,6 +443,11 @@ public class ColorimetryLiquidActivity extends BaseActivity
                     CaddisflyApp.getApp().currentTestInfo.getSwatches(),
                     maxDistance,
                     AppConfig.ColorModel.HSV));
+
+            results.add(DataHelper.analyzeColor(photoColor,
+                    CaddisflyApp.getApp().currentTestInfo.getSwatches(),
+                    maxDistance,
+                    AppConfig.ColorModel.LAB));
         }
 
         results.add(0, DataHelper.analyzeColor(photoColor,