+++ caddisfly-app/app/src/main/java/org/akvo/caddisfly/ui/OtherPreferenceFragment.java
@@ -28,7 +28,6 @@ import android.view.ViewGroup;
 import android.widget.ListView;
 
 import org.akvo.caddisfly.R;
-import org.akvo.caddisfly.app.CaddisflyApp;
 import org.akvo.caddisfly.util.ListViewUtils;
 import org.akvo.caddisfly.util.UpdateCheckTask;
 
@@ -57,7 +56,7 @@ public class OtherPreferenceFragment extends PreferenceFragment {
         if (updatePreference != null) {
             updatePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                 public boolean onPreferenceClick(Preference preference) {
-                    UpdateCheckTask updateCheckTask = new UpdateCheckTask(getActivity(), false, CaddisflyApp.getVersion(getActivity()));
+                    UpdateCheckTask updateCheckTask = new UpdateCheckTask(getActivity(), false);
                     updateCheckTask.execute();
                     return true;
                 }